Revert "move sciml code to separate files and resolve circular imports"
This reverts commit 153289ffd789f9e0a7baebaac86214a40a9d7b90.
